requirements.txt
asgiref==3.6.0
astroid==2.15.2
autopep8==2.0.1
bcrypt==4.0.1
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
lipsum==0.1.2
mccabe==0.7.0
Naked==0.1.32
platformdirs==3.2.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pylint==2.17.2
python-decouple==3.8
pytz==2023.2
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.14
webencodings==0.5.1
wrapt==1.15.0
